Split embedded iroh bridge wrapper
